FT: Centralize model version tracking for BucketInfo
Prior to this commit, we were  manually updating the
BucketInfo model version a number of places in the code
where BucketInfo is instantiated. This is likely to be
forgotten when changes are made to the BucketInfo model.
So here, we have created a static function on the BucketInfo
class which pulls the current model version and adds a changelog
 to keep track of the versions.